grid options file
