+++ src/components/creator/Editor/SceneEdition/GridOptions/GridOptions.tsx
@@ -1,0 +1,11 @@
+import { Stack } from "@mui/material";
+import { SizeEditor, StyleGridProps } from "./SizeEditor";
+
+export const GridOptions = (props: StyleGridProps) => {
+
+    return (
+        <Stack direction="row" alignItems="center" justifyContent="space-between">
+            <SizeEditor setStyleGrid={props.setStyleGrid}/>
+        </Stack>
+    )
+}
